+++ src/syntax/utils.rs
@@ -75,3 +75,5 @@ pub fn parse_token_type(
         None => Err(ParsingError::Unmatched),
     }
 }
+
+